+++ concordium-node/src/plugins/consensus.rs
@@ -8,7 +8,9 @@ use crate::{
     connection::ConnChange,
     consensus_ffi::{
         catch_up::{PeerList, PeerStatus},
-        consensus::{ConsensusContainer, ConsensusRuntimeParameters, CALLBACK_QUEUE},
+        consensus::{
+            ConsensusContainer, ConsensusRuntimeParameters, SemaphoredMessage, CALLBACK_QUEUE,
+        },
         ffi::{self, ExecuteBlockCallback, StartConsensusConfig},
         helpers::{
             ConsensusFfiResponse,
@@ -101,6 +103,39 @@ pub fn get_baker_data(
     Ok((genesis_data, private_data))
 }
 
+/// Returns true if the message is a catch-up response message.
+/// See: Serialization of a 'VersionedCatchUpStatus' message.
+/// https://github.com/Concordium/concordium-node/blob/main/concordium-consensus/src/Concordium/Types/CatchUp.hs#L24
+pub fn is_catch_up_response_message(request: &ConsensusMessage) -> anyhow::Result<bool> {
+    if request.variant != PacketType::CatchUpStatus {
+        return Ok(false);
+    }
+
+    let (discriminator, _payload) = request
+        .payload
+        .split_at_checked(7)
+        .context("Catch-up message payload should have at least 7 bytes")?;
+
+    let version = discriminator[5];
+    let tag = discriminator[6];
+
+    // Enforce valid version
+    if version != 0 && version != 1 {
+        anyhow::bail!("Invalid catch-up message version: {}", version);
+    }
+
+    // Version 0/1, use tag bytes 2 or 3 for response messages.
+    // Version 0 ("no genesis"), use tag byte 6 for response messages.
+    //
+    // This means any tag with the bit 1 set, is a response message.
+    // That is, tag AND 0b10 equals 0b10.
+    //
+    // 2 → 0b10
+    // 3 → 0b11
+    // 6 → 0b110
+    Ok(tag & 0x02 == 0x02)
+}
+
 /// Handles packets coming from other peers.
 pub fn handle_pkt_out(
     node: &P2PNode,
@@ -108,6 +143,7 @@ pub fn handle_pkt_out(
     peer_id: RemotePeerId, // id of the peer that sent the message.
     msg: Vec<u8>,
     is_broadcast: bool,
+    pending_messages_semaphore: &Arc<tokio::sync::Semaphore>,
 ) -> anyhow::Result<()> {
     let (consensus_type_bytes, payload) = msg
         .split_at_checked(1)
@@ -124,49 +160,103 @@ pub fn handle_pkt_out(
     // length of the actual payload. The message has a 1-byte tag prepended to it.
     let payload_len = payload.len();
 
-    let request = ConsensusMessage::new(
+    let consensus_message = ConsensusMessage::new(
         MessageType::Inbound(peer_id, distribution_mode),
         packet_type,
         Arc::from(msg),
         dont_relay_to,
         None,
     );
+    // TODO: Temporal logging
+    info!("SEMAPHORE VALUE:{:?}", pending_messages_semaphore);
+    let request = SemaphoredMessage::new(
+        consensus_message.clone(),
+        pending_messages_semaphore.clone().try_acquire_owned()?,
+    );
 
-    if packet_type == PacketType::Transaction {
-        if payload_len > configuration::PROTOCOL_MAX_TRANSACTION_SIZE {
-            bail!(
-                "Transaction size exceeds {} bytes.",
-                configuration::PROTOCOL_MAX_TRANSACTION_SIZE
-            )
-        }
-        if let Err(e) = CALLBACK_QUEUE.send_in_low_priority_message(request) {
-            match e.downcast::<TrySendError<QueueMsg<ConsensusMessage>>>()? {
-                TrySendError::Full(_) => {
-                    node.stats
-                        .received_consensus_messages
-                        .with_label_values(&[packet_type.label(), "dropped"])
-                        .inc();
-                    node.bad_events.inc_dropped_low_queue(peer_id);
-                }
-                TrySendError::Disconnected(_) => {
-                    panic!("Low priority consensus queue has been shutdown!")
+    match packet_type {
+        PacketType::Transaction => {
+            if payload_len > configuration::PROTOCOL_MAX_TRANSACTION_SIZE {
+                bail!(
+                    "Transaction size exceeds {} bytes.",
+                    configuration::PROTOCOL_MAX_TRANSACTION_SIZE
+                )
+            }
+            if let Err(e) = CALLBACK_QUEUE.send_in_low_priority_message(request) {
+                match e.downcast::<TrySendError<QueueMsg<ConsensusMessage>>>()? {
+                    TrySendError::Full(_) => {
+                        node.stats
+                            .received_consensus_messages
+                            .with_label_values(&[packet_type.label(), "dropped"])
+                            .inc();
+                        node.bad_events.inc_dropped_low_queue(peer_id);
+                    }
+                    TrySendError::Disconnected(_) => {
+                        panic!("Low priority consensus queue has been shutdown!")
+                    }
                 }
             }
         }
-    } else {
-        // high priority message
-        if let Err(e) = CALLBACK_QUEUE.send_in_high_priority_message(request) {
-            match e.downcast::<TrySendError<QueueMsg<ConsensusMessage>>>()? {
-                TrySendError::Full(_) => {
-                    node.stats
-                        .received_consensus_messages
-                        .with_label_values(&[packet_type.label(), "dropped"])
-                        .inc();
+        PacketType::CatchUpStatus => {
+            if is_catch_up_response_message(&consensus_message)? {
+                // Send responses if previously requested into the high priority queue.
+                // These messages need to be processed in the high priority queue synchronously to avoid race-conditions.
+                // Otherwise, a new block received via regular consensus messages
+                // may also be processed by the background queue if it was part of
+                // a catch-up response, which can lead to conflicts.
+                if let Err(e) = CALLBACK_QUEUE.send_in_high_priority_message(request) {
+                    match e.downcast::<TrySendError<QueueMsg<SemaphoredMessage>>>()? {
+                        TrySendError::Full(_) => {
+                            node.stats
+                                .received_consensus_messages
+                                .with_label_values(&[packet_type.label(), "dropped"])
+                                .inc();
 
-                    node.bad_events.inc_dropped_high_queue(peer_id);
+                            node.bad_events.inc_dropped_high_queue(peer_id);
+                        }
+                        TrySendError::Disconnected(_) => {
+                            panic!("High priority consensus queue has been shutdown!")
+                        }
+                    }
                 }
-                TrySendError::Disconnected(_) => {
-                    panic!("High priority consensus queue has been shutdown!")
+            } else {
+                // Handle catch-up message in the background queue, since it
+                // does not require the block state lock.
+                if let Err(e) = CALLBACK_QUEUE.send_in_background_message(request) {
+                    match e.downcast::<TrySendError<QueueMsg<SemaphoredMessage>>>()? {
+                        TrySendError::Full(QueueMsg::Relay(_request)) => {
+                            node.stats
+                                .received_consensus_messages
+                                .with_label_values(&[packet_type.label(), "dropped"])
+                                .inc();
+                            node.bad_events.inc_dropped_background_queue(peer_id);
+                        }
+                        TrySendError::Full(QueueMsg::Stop) => {
+                            // Should not happen, as we are not sending Stop messages here.
+                            unreachable!();
+                        }
+                        TrySendError::Disconnected(_) => {
+                            panic!("Background consensus queue has been shutdown!")
+                        }
+                    }
+                }
+            }
+        }
+        _ => {
+            // high priority message
+            if let Err(e) = CALLBACK_QUEUE.send_in_high_priority_message(request) {
+                match e.downcast::<TrySendError<QueueMsg<ConsensusMessage>>>()? {
+                    TrySendError::Full(_) => {
+                        node.stats
+                            .received_consensus_messages
+                            .with_label_values(&[packet_type.label(), "dropped"])
+                            .inc();
+
+                        node.bad_events.inc_dropped_high_queue(peer_id);
+                    }
+                    TrySendError::Disconnected(_) => {
+                        panic!("High priority consensus queue has been shutdown!")
+                    }
                 }
             }
         }